feat: Implement free money/compensation
